+++ carp_mobile_sensing/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.40.8
+
+* adding `pause` and `remove` life cycle methods to the `DeploymentController`
+* support for deleting cached runtime information (deployment and task list)
+
 ## 0.40.7
 
 * Fix of `BufferingUserAccelerometerProbe` error.